+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 All notable changes to this project will be documented in this file.
  
 
+
+## 2024-07-10
+
+## Changed
+
+- Messaging for open beta data caps centralized to krixik/utilities/validators/data/__init__.py
+- Rounding messaging updated
+
+
+
 ## Version update [1.1.18] 2024-07-09